feat: move all files
for cleaner git history
