+++ frontends/dashboard/src/analytics/api/analytics.ts
