+++ frontends/dashboard/src/analytics/api/ctr.ts
