+++ frontends/dashboard/src/analytics/api/tables.ts
