+++ frontends/dashboard/src/analytics/api/trends.ts
