+++ frontends/dashboard/src/analytics/components/CTRInfoPanel.tsx
