+++ frontends/dashboard/src/analytics/components/FilterBar.tsx
