+++ frontends/dashboard/src/analytics/components/PaginationButtons.tsx
