+++ frontends/dashboard/src/analytics/components/QueryStringDisplay.tsx
