+++ frontends/dashboard/src/analytics/components/Sidebar.tsx
