+++ frontends/dashboard/src/analytics/components/SimpleTimeRangeSelector.tsx
