+++ frontends/dashboard/src/analytics/components/SingleQueryInfo/DataSquare.tsx
