+++ frontends/dashboard/src/analytics/components/SingleQueryInfo/ResultCard.tsx
