+++ frontends/dashboard/src/analytics/components/charts/CTRSummary.tsx
