+++ frontends/dashboard/src/analytics/components/charts/Card.tsx
