+++ frontends/dashboard/src/analytics/components/charts/HeadQueries.tsx
