+++ frontends/dashboard/src/analytics/components/charts/LatencyGraph.tsx
