+++ frontends/dashboard/src/analytics/components/charts/LowConfidenceQueries.tsx
