+++ frontends/dashboard/src/analytics/components/charts/NoResultQueries.tsx
