+++ frontends/dashboard/src/analytics/components/charts/QueryCounts.tsx
