+++ frontends/dashboard/src/analytics/components/charts/RAGUsageGraph.tsx
