+++ frontends/dashboard/src/analytics/components/charts/RagQueries.tsx
