+++ frontends/dashboard/src/analytics/components/charts/SearchUsageGraph.tsx
