+++ frontends/dashboard/src/analytics/contexts/UserAuthContext.tsx
