+++ frontends/dashboard/src/analytics/hooks/data/useDataExplorerSearch.ts
@@ -3,7 +3,7 @@ import { getSearchQueries } from "../../api/tables";
 import { createEffect, createSignal, useContext } from "solid-js";
 import { createStore } from "solid-js/store";
 import { subDays } from "date-fns";
-import { usePagination } from "./../usePagination";
+import { usePagination } from "../usePagination";
 import { DatasetContext } from "../../layouts/TopBarLayout";
 import { AnalyticsParams } from "shared/types";
