+++ frontends/dashboard/src/analytics/hooks/useCTRNeedsSetup.ts
