+++ frontends/dashboard/src/analytics/hooks/usePagination.ts
