+++ frontends/dashboard/src/analytics/hooks/usePathname.ts
