+++ frontends/dashboard/src/analytics/index.css
