+++ frontends/dashboard/src/analytics/index.tsx
