+++ frontends/dashboard/src/analytics/layouts/DataExplorerTabs.tsx
