+++ frontends/dashboard/src/analytics/layouts/TopBarLayout.tsx
