+++ frontends/dashboard/src/analytics/pages/CTRAnalyticsPage.tsx
