+++ frontends/dashboard/src/analytics/pages/RagAnalyticsPage.tsx
