+++ frontends/dashboard/src/analytics/pages/SearchAnalyticsPage.tsx
