+++ frontends/dashboard/src/analytics/pages/SingleQueryPage.tsx
