+++ frontends/dashboard/src/analytics/pages/TrendExplorer.tsx
