+++ frontends/dashboard/src/analytics/pages/tablePages/RAGTablePage.tsx
