+++ frontends/dashboard/src/analytics/pages/tablePages/SearchTablePage.tsx
