+++ frontends/dashboard/src/analytics/types.d.ts
