+++ frontends/dashboard/src/analytics/utils/searchType.ts
