+++ frontends/dashboard/src/analytics/utils/titleCase.ts
