+++ frontends/dashboard/src/analytics/utils/useBetterNav.ts
