+++ frontends/dashboard/src/analytics/vite-env.d.ts
